:recycle: Move Associativity into its own module
