+++ zia/src/context_snap_shot.rs
@@ -30,10 +30,7 @@ use generic_array::{
     arr, functional::FunctionalSequence, ArrayLength, GenericArray,
 };
 use maplit::hashset;
-use std::{
-    collections::{HashMap, HashSet, VecDeque},
-    fmt::Display,
-};
+use std::collections::{HashMap, HashSet, VecDeque};
 
 type ConceptId = usize;
 
@@ -49,54 +46,6 @@ pub struct ContextSnapShot {
     /// Keeps track of indices of the `concepts` field that have `None`.
     gaps: Vec<ConceptId>,
     concrete_concepts: BiMap<ConceptId, ConcreteConceptType>,
-}
-
-#[derive(Debug, PartialEq)]
-pub enum Associativity {
-    Left,
-    Right,
-}
-
-impl Associativity {
-    pub fn display_joint_left(
-        &self,
-        leftleft: impl Display,
-        leftright: impl Display,
-    ) -> String {
-        match &self {
-            Self::Left => format!("{} {}", leftleft, leftright),
-            Self::Right => format!("({} {})", leftleft, leftright),
-        }
-    }
-
-    pub fn display_joint_right(
-        &self,
-        rightleft: impl Display,
-        rightright: impl Display,
-    ) -> String {
-        match &self {
-            Self::Left => format!("({} {})", rightleft, rightright),
-            Self::Right => format!("{} {}", rightleft, rightright),
-        }
-    }
-
-    pub fn slice_tokens<'a>(
-        &self,
-        tokens: &'a [String],
-        prev_lp_index: Option<usize>,
-        lp_index: usize,
-    ) -> &'a [String] {
-        match &self {
-            Self::Left => match prev_lp_index {
-                Some(i) => &tokens[i..lp_index],
-                None => &tokens[..lp_index],
-            },
-            Self::Right => match prev_lp_index {
-                Some(i) => &tokens[lp_index..i],
-                None => &tokens[lp_index..],
-            },
-        }
-    }
 }
 
 impl ContextSnapShot {
